hms_v2

+++ Back-End/src/main/java/in/cdac/hms/model/Role.java
@@ -20,10 +20,10 @@ import lombok.Setter;
 @Table(name = "roles")
 public class Role {
 	
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	@Id	
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	private Integer id;
+	
 	@Column(nullable = false, unique = true)
-    private String name;
-   
+    private String name;   
 }
