+++ Back-End/src/main/java/in/cdac/hms/security/UserPrincipal.java
@@ -16,11 +16,8 @@ import java.util.stream.Collectors;
 public class UserPrincipal implements UserDetails {
     
 	private static final long serialVersionUID = 1L;
-
-	private Long id;
-
+	private Integer id;
     private String name;
-
     private String username;
 
     @JsonIgnore
@@ -31,7 +28,7 @@ public class UserPrincipal implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserPrincipal(Long id, String name, String username, String email, String password, Collection<? extends GrantedAuthority> authorities) {
+    public UserPrincipal(Integer id, String name, String username, String email, String password, Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.name = name;
         this.username = username;
@@ -55,7 +52,7 @@ public class UserPrincipal implements UserDetails {
         );
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
